requirements.txt
openai
fastapi
requests
pydantic
redis
uvicorn
python-magic
unstructured
python-telegram-bot
python-dotenv